requirements.txt
streamlit
PyPDF2
pdf2image
pillow
pytesseract
numpy
faiss-cpu
sentence-transformers
spacy
groq
